+++ src/mnextend/__init__.py
@@ -2,10 +2,14 @@
 #
 # License: BSD (3-clause)
 
+from importlib.metadata import version
+
 from mnextend.iclabel import IC_LABELS, plot_ica_components, run_iclabel
 from mnextend.io.readers import read_epochs, read_raw
 from mnextend.io.utils import split_name_ext
 from mnextend.io.writers import write_epochs, write_raw
+
+__version__ = version("mnextend")
 
 __all__ = [
     "IC_LABELS",