+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [UNRELEASED] · YYYY-MM-DD
+### ✨ Added
+- Add `__version__` attribute to the package ([#5](https://github.com/cbrnr/mnextend/pull/5) by [Clemens Brunner](https://github.com/cbrnr))
 
 ## [0.2.0] · 2026-06-25
 ### ✨ Added